+++ src/osu/versions/no_leniency/mod.rs
@@ -280,7 +280,7 @@ mod tests {
     #[test]
     #[ignore]
     fn no_leniency_single() {
-        let file = match File::open("./maps/1791963.osu") {
+        let file = match File::open("./maps/2573164.osu") {
             Ok(file) => file,
             Err(why) => panic!("Could not open file: {}", why),
         };
@@ -291,11 +291,11 @@ mod tests {
         };
 
         let result = OsuPP::new(&map)
-            .n300(1206)
-            .n100(15)
-            .n50(0)
-            .combo(1643)
-            .mods(24)
+            .mods(0)
+            // .n300(1206)
+            // .n100(15)
+            // .n50(0)
+            // .combo(1643)
             .calculate();
 
         println!("Stars: {}", result.stars());